requirements.txt
# Hydromind-Demo 依赖清单
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0

# 空间数据处理（分布式模型扩展）
# rasterio>=1.3.0
# geopandas>=0.14.0
